src/requirements.freeze.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
PyYAML==3.10
Unidecode==0.04.14
Werkzeug==0.9.4
argparse==1.2.1
coverage==3.7.1
itsdangerous==0.23
mailer==0.7
nose==1.3.0
nosexcover==1.0.8
pyaml==13.12.0
pymongo==2.6.3
requests==2.2.1
wsgiref==0.1.2
singleton==0.1.0
Flask-SQLAlchemy==1.0
SQLAlchemy==0.8.2